outsource EQA document types and annotations to documents package
