+++ src/pie_models/documents/__init__.py
@@ -1,0 +1,2 @@
+from .annotation import ExtractiveAnswer, Question
+from .document import ExtractiveQADocument, TokenizedExtractiveQADocument
